Refactor tests to use a generic check for which variant is needed.
