+++ tests/component_tests/epaper_spi/test_init.py
@@ -13,12 +13,7 @@ from esphome.components.epaper_spi.display import (
     FINAL_VALIDATE_SCHEMA,
     MODELS,
 )
-from esphome.components.esp32 import (
-    KEY_BOARD,
-    KEY_VARIANT,
-    VARIANT_ESP32,
-    VARIANT_ESP32S3,
-)
+from esphome.components.esp32 import KEY_BOARD, KEY_VARIANT, VARIANT_ESP32
 from esphome.const import (
     CONF_BUSY_PIN,
     CONF_CS_PIN,
@@ -33,6 +28,30 @@ from esphome.const import (
 )
 from esphome.types import ConfigType
 from tests.component_tests.types import SetCoreConfigCallable
+
+# Pin options whose values must be valid on the chosen ESP32 variant.
+_PIN_CONF_KEYS = (
+    CONF_CS_PIN,
+    CONF_DC_PIN,
+    CONF_RESET_PIN,
+    CONF_BUSY_PIN,
+    CONF_ENABLE_PIN,
+)
+
+
+def _pins_for(model: Any, config: ConfigType) -> list:
+    """Collect every GPIO the config will actually use (model defaults or injected)."""
+    pins: list = []
+    for key in _PIN_CONF_KEYS:
+        # An injected value in the config takes precedence over the model default.
+        value = config[key] if key in config else model.get_default(key)
+        if not value:  # get_default returns False for pins the model omits
+            continue
+        if isinstance(value, list):
+            pins.extend(value)
+        else:
+            pins.append(value)
+    return pins
 
 
 def run_schema_validation(
@@ -93,29 +112,20 @@ def test_basic_configuration_errors(
 def test_all_predefined_models(
     set_core_config: SetCoreConfigCallable,
     set_component_config: Callable[[str, Any], None],
+    choose_variant_with_pins: Callable[[list], None],
 ) -> None:
     """Test all predefined epaper models validate successfully with appropriate defaults."""
 
+    set_core_config(
+        PlatformFramework.ESP32_IDF,
+        platform_data={KEY_BOARD: "esp32dev", KEY_VARIANT: VARIANT_ESP32},
+    )
+
+    # Configure SPI component which is required by epaper_spi
+    set_component_config("spi", {"id": "spi_bus", "clk_pin": 18, "mosi_pin": 19})
+
     # Test all models, providing default values where necessary
     for name, model in MODELS.items():
-        # SEEED models are designed for ESP32-S3 hardware
-        if name in ("SEEED-EE04-MONO-4.26", "SEEED-RETERMINAL-E1002"):
-            set_core_config(
-                PlatformFramework.ESP32_IDF,
-                platform_data={
-                    KEY_BOARD: "esp32-s3-devkitc-1",
-                    KEY_VARIANT: VARIANT_ESP32S3,
-                },
-            )
-        else:
-            set_core_config(
-                PlatformFramework.ESP32_IDF,
-                platform_data={KEY_BOARD: "esp32dev", KEY_VARIANT: VARIANT_ESP32},
-            )
-
-        # Configure SPI component which is required by epaper_spi
-        set_component_config("spi", {"id": "spi_bus", "clk_pin": 18, "mosi_pin": 19})
-
         config = {"model": name}
 
         # Add ID field
@@ -144,6 +154,10 @@ def test_all_predefined_models(
         if not model.get_default(CONF_CS_PIN):
             config[CONF_CS_PIN] = 5
 
+        # Select an ESP32 variant on which all of this model's pins are valid
+        # (some models default to high-numbered pins only present on the S3).
+        choose_variant_with_pins(_pins_for(model, config))
+
         run_schema_validation(config)
 
 
@@ -155,27 +169,19 @@ def test_individual_models(
     model_name: str,
     set_core_config: SetCoreConfigCallable,
     set_component_config: Callable[[str, Any], None],
+    choose_variant_with_pins: Callable[[list], None],
 ) -> None:
     """Test each epaper model individually to ensure it validates correctly."""
-    # SEEED models are designed for ESP32-S3 hardware
-    if model_name in ("SEEED-EE04-MONO-4.26", "SEEED-RETERMINAL-E1002"):
-        set_core_config(
-            PlatformFramework.ESP32_IDF,
-            platform_data={
-                KEY_BOARD: "esp32-s3-devkitc-1",
-                KEY_VARIANT: VARIANT_ESP32S3,
-            },
-        )
-    else:
-        set_core_config(
-            PlatformFramework.ESP32_IDF,
-            platform_data={KEY_BOARD: "esp32dev", KEY_VARIANT: VARIANT_ESP32},
-        )
+    model = MODELS[model_name]
+
+    set_core_config(
+        PlatformFramework.ESP32_IDF,
+        platform_data={KEY_BOARD: "esp32dev", KEY_VARIANT: VARIANT_ESP32},
+    )
 
     # Configure SPI component which is required by epaper_spi
     set_component_config("spi", {"id": "spi_bus", "clk_pin": 18, "mosi_pin": 19})
 
-    model = MODELS[model_name]
     config: dict[str, Any] = {"model": model_name, "id": "test_display"}
 
     # Add required fields based on model defaults
@@ -197,6 +203,10 @@ def test_individual_models(
 
     if not model.get_default(CONF_CS_PIN):
         config[CONF_CS_PIN] = 5
+
+    # Select an ESP32 variant on which all of this model's pins are valid
+    # (some models default to high-numbered pins only present on the S3).
+    choose_variant_with_pins(_pins_for(model, config))
 
     # This should not raise any exceptions
     run_schema_validation(config)
